parse funcitonality is settled, fragment in flux
